Refactor hammerspoon and nvim configurations
This commit includes significant changes to both the hammerspoon and
neovim configurations. In hammerspoon, the hotkey bindings for opening
applications have been refactored into inline functions for better
readability and compactness. New logic has also been introduced to
handle marking of applications for later focus.

In the neovim configuration, we have similarly restructured the vim
options for increased readability. In addition, the Lua-based plugin
configurations have been overhauled, with the addition of a new
refactoring plugin and other modifications to improve development
utility.

A new asynchronous command for "rerun last" has also been introduced,
along with enhancements to multiple cursors usage, refactoring
functionality and other usability improvements.

In addition, there are changes in the plugins we are using and various
other settings adjustments and enhancements that help to make the
development process smoother and more efficient.